Feature: Split advanced-operations into operations/* #577
